+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   which is slow compared to checking for equality or pattern matching (#2180).
 - The new `gleam remove <package_name>` can be used to remove dependencies
   from a Gleam project.
+- The `gleam docs build` command gains the `--open` flag to open the docs after they are generated (#2188).
 
 ## v0.29.0 - 2023-05-23